requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]
python-multipart
email-validator
httpx
bcrypt==4.0.1
python-jose[cryptography]
requests>=2.0.0